requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiofiles==23.2.1
apscheduler==3.10.4
nest-asyncio==1.5.8
flask==3.0.0